FatFS sample code 20131116
  Jun 22, 2013  AVR/PIC24: Updated MMC drivers.
  Jul 28, 2013  Added a foolproof project for ATMEL AVR.
  Oct 02, 2013  Modified for FatFs module R0.10.
  Nov 16, 2013  MMC modules: Fixed CMD12 to stop data transmisson can fail.

From http://elm-chan.org/fsw/ff/ffsample.zip

386a4810fc259125251297d8a9a330c2  ffsample.zip

+++ lpc23xx/diskio.c
@@ -39,7 +39,7 @@ DRESULT disk_read (
 	BYTE drv,
 	BYTE* buff,
 	DWORD lba,
-	BYTE count
+	UINT count
 )
 {
 	switch (drv) {
@@ -56,7 +56,7 @@ DRESULT disk_write (
 	BYTE drv,
 	const BYTE* buff,
 	DWORD lba,
-	BYTE count
+	UINT count
 )
 {
 	switch (drv) {
